Moved canSortResource Mixin
